refactor: Remove duplication code
